# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Leggen API",
...
    "version": "dev"
  },
  "paths": {
...
    "/api/v1/category-rules/reference": {
      "get": {
        "tags": [
          "category-rules"
        ],
        "summary": "Get Rule Reference",
        "description": "The rule-scripting reference: every `tx` field, every stdlib function,\nthe sandbox rules, and example scripts.\n\nFetch this before writing a rule. It is generated from the runtime's own\ndefinitions, so it is always current for this server.",
        "operationId": "get_rule_reference_api_v1_category_rules_reference_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleReference"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/category-rules/test": {
      "post": {
        "tags": [
          "category-rules"
        ],
        "summary": "Test Rule Script",
        "description": "Run a script against one existing transaction and see the result and\nits log() output.\n\nRejects a script that does not compile with 422; a runtime error inside\nthe script comes back in `error` with a 200.",
        "operationId": "test_rule_script_api_v1_category_rules_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleScriptTest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleTestResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/category-rules/preview": {
      "post": {
        "tags": [
          "category-rules"
        ],
        "summary": "Preview Rule Script",
        "description": "List every transaction a script matches, without saving anything.\n\nManually categorized transactions are included and flagged `manual`:\nthe engine would leave them alone, but they show the rule's true reach.\nUse this to check a rule's precision before creating or activating it.",
        "operationId": "preview_rule_script_api_v1_category_rules_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleScriptPreview"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulePreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/category-rules/apply": {
      "post": {
        "tags": [
          "category-rules"
        ],
        "summary": "Apply Rules",
        "description": "Run every active rule over all transactions without a manual category.\n\nRules also run automatically over new transactions on every sync; this\nre-evaluates history, which is what a new or edited rule needs. Manual\nassignments are never changed. Rule-made assignments no active rule\nreproduces any more are cleared.",
        "operationId": "apply_rules_api_v1_category_rules_apply_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "dry_run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Report what would change without writing anything.",
              "default": false,
              "title": "Dry Run"
            },
            "description": "Report what would change without writing anything."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleRunReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/category-rules": {
      "get": {
        "tags": [
          "category-rules"
        ],
        "summary": "List Rules",
        "description": "All rules in evaluation order: ascending priority, then id.",
        "operationId": "list_rules_api_v1_category_rules_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryRule"
                  },
                  "type": "array",
                  "title": "Response List Rules Api V1 Category Rules Get"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      },
      "post": {
        "tags": [
          "category-rules"
        ],
        "summary": "Create Rule",
        "description": "Create a rule. The script is compiled first and rejected with 422 if\nit has a syntax error. Creating a rule does not run it: call\nPOST /category-rules/apply, or wait for the next sync to categorize new\ntransactions.",
        "operationId": "create_rule_api_v1_category_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryRuleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRule"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          },
          "409": {
            "description": "A rule with this name already exists"
          },
          "422": {
            "description": "The script does not compile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/category-rules/{rule_id}": {
      "get": {
        "tags": [
          "category-rules"
        ],
        "summary": "Get Rule",
        "description": "One rule by id.",
        "operationId": "get_rule_api_v1_category_rules__rule_id__get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRule"
                }
              }
            }
          },
          "404": {
            "description": "Rule not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "category-rules"
        ],
        "summary": "Update Rule",
        "description": "Update the fields sent; the others keep their value. A changed script\nis compiled first. Editing does not re-run the engine.",
        "operationId": "update_rule_api_v1_category_rules__rule_id__put",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryRuleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRule"
                }
              }
            }
          },
          "404": {
            "description": "Rule or category not found"
          },
          "409": {
            "description": "A rule with this name already exists"
          },
          "422": {
            "description": "The script does not compile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "category-rules"
        ],
        "summary": "Delete Rule",
        "description": "Delete a rule. The categories it assigned are removed with it;\nmanual assignments are untouched.",
        "operationId": "delete_rule_api_v1_category_rules__rule_id__delete",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Rule not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RulePreviewResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/RuleMatch"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Transactions matched."
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_prev": {
            "type": "boolean",
            "title": "Has Prev"
          },
          "evaluated": {
            "type": "integer",
            "title": "Evaluated",
            "description": "Transactions the script was run against."
          },
          "errors": {
            "type": "integer",
            "title": "Errors",
            "description": "Evaluations that raised a runtime error."
          },
          "error_samples": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Error Samples"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "per_page",
          "total_pages",
          "has_next",
          "has_prev",
          "evaluated",
          "errors"
        ],
        "title": "RulePreviewResponse",
        "description": "The transactions a script matches, paginated, with run statistics."
      },
      "RuleReference": {
        "properties": {
          "stdlib": {
            "items": {
              "$ref": "#/components/schemas/RuleFunctionGroup"
            },
            "type": "array",
            "title": "Stdlib"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/RuleFieldGroup"
            },
            "type": "array",
            "title": "Fields"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "max_instructions": {
            "type": "integer",
            "title": "Max Instructions"
          },
          "examples": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Examples",
            "description": "Small, complete scripts showing common shapes of rule."
          }
        },
        "type": "object",
        "required": [
          "stdlib",
          "fields",
          "notes",
          "max_instructions",
          "examples"
        ],
        "title": "RuleReference",
        "description": "Everything a script can use: the `tx` fields and the stdlib."
      },
...
      "RuleScriptPreview": {
        "properties": {
          "lua_script": {
            "type": "string",
            "minLength": 1,
            "title": "Lua Script",
            "description": "Body of a Lua function receiving the transaction as `tx`; return true to apply the rule. See GET /category-rules/reference for the fields and functions available."
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "default": 1
          },
          "per_page": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Per Page",
            "default": 50
          }
        },
        "type": "object",
        "required": [
          "lua_script"
        ],
        "title": "RuleScriptPreview",
        "description": "Find every transaction a script matches."
      },
      "RuleScriptTest": {
        "properties": {
          "lua_script": {
            "type": "string",
            "minLength": 1,
            "title": "Lua Script",
            "description": "Body of a Lua function receiving the transaction as `tx`; return true to apply the rule. See GET /category-rules/reference for the fields and functions available."
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          }
        },
        "type": "object",
        "required": [
          "lua_script",
          "account_id",
          "transaction_id"
        ],
        "title": "RuleScriptTest",
        "description": "Evaluate a script against one existing transaction."
      },
      "RuleTestResult": {
        "properties": {
          "matched": {
            "type": "boolean",
            "title": "Matched"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Runtime error raised by the script, if any. A script that does not compile is rejected with 422 instead."
          },
          "logs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Logs",
            "description": "Lines the script passed to log()."
          }
        },
        "type": "object",
        "required": [
          "matched"
        ],
        "title": "RuleTestResult",
        "description": "Outcome of evaluating a script against one transaction."
      },
...
      "ErrorResponse": {
        "description": "Unified error envelope returned by every API error response.\n\n`detail` is always a plain string, on every status code including 422.\nClients may rely on that: it is the backward-compatible contract with\nthe frontend (`getApiErrorMessage`) and the CLI (`api_client.py`).\n\nAn optional `request_id` could be added here later without breaking\nclients, should log correlation ever become worth the plumbing.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorField"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          }
        },
        "required": [
          "detail",
          "code",
          "status"
        ],
        "title": "ErrorResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}